Extract Month Converter to new class.
